pyproject.toml
[project]
name = "aida"
version = "0.1.0"
description = "Active Learning Framework for Object Detection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.9.4",
    "seaborn>=0.13.2",
    "pandas==2.2.3",
    "ultralytics>=8.3.179",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "PyYAML>=6.0",
    "scikit-learn>=1.3.0",
    "scikit-multilearn>=0.2.0",
    "clip",
    "torch==2.3.0",
    "torchvision==0.18.0",
    "markupsafe>=2.1.1,<3.0",
    "loguru>=0.7.3",
    "wandb>=0.21.3",
    "dotenv>=0.9.9",
    "kaggle>=1.7.4.5",
    "rich>=14.1.0",
    "setproctitle>=1.3.7",
    "setuptools>=70",
    "wheel",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
clip = { git = "https://github.com/ultralytics/CLIP.git" }
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"

[project.scripts]
aida-strategy = "scripts.strategy:main"
aida-train = "scripts.train:main"
aida-val = "scripts.val:main"
aida-experiment = "scripts.run_experiment:main"
